requirements.txt
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
pandas==2.1.3
openpyxl==3.1.2
ollama==0.1.6
python-dateutil==2.8.2
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0